Dodanie endpointa do pozyskiwania pojedynczego kursu

+++ src/controllers/course.ts
@@ -47,7 +47,7 @@ export const updateCourse = (req: Request, res: Response) => {
 
         coursesDatabase.query(
             'UPDATE Courses SET trainer_id = ?, title = ?, language = ?, level = ?, location = ?, image_path = ?) WHERE id = ?',
-            [data.trainer, data.title, data.language, data.level, data.location, req.file?.path, courseId],
+            [data.trainer, data.title, data.language, data.level, data.location, req.file?.path ? req.file.path : originalImagePath, courseId],
             (error, result) => {
                 if (error || !result.affectedRows) {
                     if (req.file?.path)
@@ -75,13 +75,30 @@ export const updateCourse = (req: Request, res: Response) => {
                             }
                         )
                     })
-                    fs.unlink(originalImagePath, error => console.log(error));
+                    if (req.file?.path)
+                        fs.unlink(originalImagePath, error => console.log(error));
                     return res.json({
-                        message: 'Udało się dodać kurs!'
+                        message: 'Udało się zaktualizować kurs!'
                     })
 
                 })
             }
         )
+    })
+}
+
+export const getCourse = (req: Request, res: Response) => {
+    const courseId = req.params.id;
+
+    coursesDatabase.query('SELECT * FROM Courses WHERE id = ?', [courseId], (error, result) => {
+        if (error) return res.sendStatus(500);
+
+        let {id, owner_id, image_path, ...data} = result[0];
+        coursesDatabase.query('SELECT * FROM Lessons WHERE course_id = ?', [courseId], (error, result) => {
+            if (error) return res.sendStatus(500);
+
+            data = {...data, datetimes: result};
+            return res.json(data);
+        })
     })
 }
